update_ssh_config_vagrant: initial import from kdevops
Publish the update_ssh_config_vagrant kdevops role.  This is an import
from kdevops's code as of commit 459ae451097 ("ansible: move devconfig
role to ansible galaxy").

The only change here was to add the python script here and also to
change the path of when we call the script. Let's see if this works
well as a ansible-galaxy module, it's a bit of a hack given we are
changing directories into the vagrant dir. We can probably address
this better by letting the user override that. We can do that later.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>