tasks: add support for a custom kexalgorithms
We also simplify our calls by just asking the python script to
read the vagrant ssh-config output and augment it. This is
required, in order to support the custom kexalgorithms.

Another big thing that changes is that older hosts, such
as SLES10-SP3 do not support modern versions of ansible,
as it has Python 2.4 and we need at least Python 2.6.

Provisioning the ssh configuration however should require
to be able to use ansible on the target host when using vagrant,
given all we really need to do is just process the result of
the command vagrant ssh-config, and augment it, if needed.

So this role should change to run on the localhost instead,
on the controller. This means that we can remove the provisioning
aspect to update the ssh configuration from vagrant to kdevops
by just issuing a local ansible-playbook call.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>