tasks: fix ambigous use of kexalgorithms as a string
Turns out that the stupid ansible-lint complaint about not
not comparing a variable to an empty string, and *recommending*
to just use the variable alone *is not a good idea* after all.

So ansible *wants* us to use two seperate variables to make this
real clean. So let us just do that. And keep the pendantic ...
folks who write a stupid lint happy.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>